+++ test/Main.hs
@@ -7,6 +7,7 @@ import qualified AWS.EventBridge.MinutesSpec as MinutesSpec
 import qualified AWS.EventBridge.MonthsSpec as MonthsSpec
 import qualified AWS.EventBridge.YearsSpec as YearsSpec
 import qualified AWS.EventBridge.RateSpec as RateSpec
+import qualified AWS.EventBridge.OneTimeSpec as OneTimeSpec
 import Test.Tasty (defaultMain, testGroup)
 
 main :: IO ()
@@ -20,4 +21,5 @@ main =
       , MonthsSpec.tests
       , YearsSpec.tests
       , RateSpec.tests
+      , OneTimeSpec.tests
       ]